Add inode creation.
Add inode deletion.

Add ext2fs feature detection and compatibility checking.

Add 64-bit file size support ( in theory ).

Add support for directory entry type field ( only on write ).

Fix grow_by bug where it was trying to allocate the last block to make sure it exists in a non-existant indirect block,
because of a quirk in checking for the bytes used in the last block.

Fix inode write gobbling a byte at the ext2 block boundry because of an iterator.next() follwed by a break causing the value of the iterator to be discarded.

TODO: Inode deletion leaks the entry

Plan for fixing:
For the inode deletion, write a shrink_to_fit function in the distant future
(chances are nobody is gonna notice a couple of bytes leaked every now and then, besides we have enough disk space right :P /jk).

Mini-Whitepaper on inode creation, because it's the most complicated part:
First allocate the inode in the block group that the last entry in the directory is in ( this is the easy part ).
Then to create a new entry in the directory for that inode:
- First fill in the values of the new entry.
- Then calculate where the new entry will be ( the new_entry_first_byte ), to do that:
	- Calculate the minimum size of the last entry by adding the size of a directory header to the length of the last entry's string, taking into account that the new entry would be after it so the size must be a multiple of 4 to keep the new entry 4-byte aligned
	- Then calculate the free space in the last entry, by subtracting the minimum size from the actual entry size
	- Then if enough free space is found in the last entry then set the entry_size of the last entry to the minimum size and write the updated entry to disk
	- Finally, wether or not the new entry could fit in the free space of the last entry, return the last byte of the last entry, this will either put the new entry in the same block as the last entry, on a 4-byte aligned boundry, if we found space in the last entry, or in a new block, since the last entry always "points" to the end of its block.
- Then "grow" the new entry to the end of the block ( since it is now the new last entry ); take the end byte of the new entry ( which can be calculated using the new_entry_first_byte and it's current size ) modulo the size of a block, and subtract it from the size of the block, this will give you the amount to add to reach the next
multiple of the size of the block, in other words it gives you the amount to add to the entry_size to make the new entry "point" to the end of it's block.
- Then resize the directory to fit the new entry ( the new last entry ) by resizing it to the end byte of the new entry ( new_entry_first_byte + new_entry entry_size )
- Write the new entry to the directory